% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_values_by_group.R
\name{average_values_by_group}
\alias{average_values_by_group}
\title{Average Values By Group Function}
\usage{
average_values_by_group(dataset, group, measure)
}
\arguments{
\item{dataset}{is the parameter for the \strong{dataset} you are inputting into the function.}

\item{group}{is the parameter for the variable you choose to \strong{group} the dataset by. This variable should be a column name from the dataset.}

\item{measure}{is the parameter for the \strong{measure} you choose to find the mean value of for each of the groups. The column you choose for the measure parameter must have numeric values.}
}
\value{
This function returns a dataframe with each \emph{group} in the first column, and the \emph{mean value of the measure for each group} in the second column.
}
\description{
This function takes a dataset and groups it by a chosen variable, then finds the mean value of a chosen measure for each group.
}
\examples{

library(tibble)

sample_data_1 <- tibble(
  Name = c("Alice", "Bob", "Charlie", "David", "Eve"),
  Age = c(25, 32, 28, 22, 30),
  City = c("New York", "New York", "New York", "Houston", "Houston"),
  Hobby = c("Gaming", "Cooking", "Gaming", "Gaming", "Cooking"))

sample_data_2 <- tibble(
  Gender = c("Female", "Male", "Male", "Male", "Female"),
  Grade = c(94, 78, 82, NA, 88))

average_values_by_group(sample_data_1, City, Age)
average_values_by_group(sample_data_1, Hobby, Age)
average_values_by_group(sample_data_2, Gender, Grade)
}
